Fix the 19 :feature:editor compile errors on the caption block
The lane's caption work (2/4-4/4) never compiled: :feature:editor is a module the
local gate cannot build, so all 19 errors were first seen in CI. Fixed as one set.

Missing / wrong imports (the "moved call drags its imports" class):
- TextLane.kt imported TimelineMarks from com.redcut.feature.editor, but it lives
  in ...editor.timeline (this file's own package) -- drop the import. Add the
  missing EditDocument (line 86), textLaneItems (line 92, a :domain projection in
  ...editor) and awaitFirstDown (line 292) imports.
- EditorViewModel.kt:762 and TextGestureSession.kt used the textOverlayById
  extension (defined in :domain:document) with no import -- add it to both.
- TimelineDraw's TRIM_EDGE_PX was file-private; make it internal so TextLane can
  share the clip trim's edge width instead of inventing a second one.

Real type/visibility bugs:
- TextInspector.kt: DEFAULT_BACKGROUND_ARGB = 0x99000000 is inferred Long (the
  literal overflows Int), but TextSpec.backgroundArgb is Int?. Write .toInt(),
  matching SHADOW_COLOR / READOUT_COLOR in TextOverlayLayer.kt.
- TimelineCanvas.kt: rememberTimelineActions declared Pair<TimelineGestures,
  State<ReorderDrag?>> but returned the unwrapped value (it held the state with
  `by`). Hold the State OBJECT so the caller's reorderDrag.value stays reactive.
- TimelineGestures.kt: reorderDragPress wraps detectDragGesturesAfterLongPress,
  which is a PointerInputScope extension -- so its receiver is PointerInputScope,
  not AwaitPointerEventScope (that was the receiver mismatch at 420 and 353).
  textLaneTrimPress is an AwaitPointerEventScope extension that opened its OWN
  awaitEachGesture; give the call site the awaitEachGesture (trimGesture's shape)
  and drop the inner one, so the scopes nest the way the rest of the file does.

Local checks that can run on this host: ktlint (main+test), detekt, arch
boundaries, test-name and param-target checkers -- all clean. The Android
compile is CI's to prove.

+++ feature/editor/src/main/kotlin/com/redcut/feature/editor/TextInspector.kt
@@ -391,7 +391,7 @@ private val SWATCH_BORDER_SELECTED = 2.dp
  * whatever the frame holds without fully covering it. The value the preview draws and the export burns —
  * it is a spec value, not a theme colour, for the reason the swatch set is.
  */
-private const val DEFAULT_BACKGROUND_ARGB = 0x99000000
+private const val DEFAULT_BACKGROUND_ARGB = 0x99000000.toInt()
 
 private val SEGMENTED_SHAPE = RoundedCornerShape(8.dp)
 private val SEGMENTED_BORDER_WIDTH = 1.dp
